+++ NEWS.md
@@ -1,3 +1,10 @@
+# dabr 0.0.2
+
+* Added option to return tables' attributes with `dabr::list_table(quiet = TRUE)`. 
+* Added function to get the attributes of a table, `dabr::get_attr`.
+* Updated queries for `dabr::insert`, `dabr::select`, and `dabr::update`.
+* Added function to verify connection objects, `dabr::is.connected`
+
 # dabr 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.